Remove unused/obsolete code for WebView2 PDF creation (BL-11437)
