+++ src/WebView2PdfMaker/WebView2PdfMaker.csproj
@@ -62,11 +62,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ConversionProgress.cs">
+    <Compile Include="ConversionHost.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="ConversionProgress.Designer.cs">
-      <DependentUpon>ConversionProgress.cs</DependentUpon>
+    <Compile Include="ConversionHost.Designer.cs">
+      <DependentUpon>ConversionHost.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -76,8 +76,8 @@
     <Compile Include="WebView2PdfComponent.Designer.cs">
       <DependentUpon>WebView2PdfComponent.cs</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="ConversionProgress.resx">
-      <DependentUpon>ConversionProgress.cs</DependentUpon>
+    <EmbeddedResource Include="ConversionHost.resx">
+      <DependentUpon>ConversionHost.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
